fix(dongtai-agent): catch all check function to avoid thread abort;
